Refactored components/Table
